+++ fixtures/mini-spring-partner-mock-app/src/main/java/com/wina/partner/matMgmt/decoSheet/service/DecoSheetServiceImpl.java
@@ -1,0 +1,26 @@
+package com.wina.partner.matMgmt.decoSheet.service;
+
+import com.wina.partner.matMgmt.decoSheet.repository.DecoSheetRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class DecoSheetServiceImpl implements DecoSheetService {
+    @Autowired
+    private DecoSheetRepository decoSheetRepository;
+
+    @Override
+    public Object retrieveDecoShetAbcsRqusList() {
+        return decoSheetRepository.retrieveDecoShetAbcsRqusList();
+    }
+
+    @Override
+    public Object insertTbHods401() {
+        return decoSheetRepository.insertTbHods401();
+    }
+
+    @Override
+    public Object insertTbHods401B() {
+        return decoSheetRepository.insertTbHods401B();
+    }
+}